+++ Additionals/ChangeLog.md
@@ -5,5 +5,6 @@
 * Templates and samples saved in settings
 * Detection mode is automatically disabled when then samples tab is deactivated
 * Background color of detection mode changed
-* Clip types updated if clip is removed
 * Scoreboard contents shown in title bar
+* Clip selected if clip property is changed
+* Clip types updated if clip is removed